convertDF {iMRMC}	R Documentation

Convert MRMC data frames

Description

Convert MRMC data frames

Usage

convertDF(inDF, inDFtype, outDFtype, readers)

Arguments

inDF
An MRMC dataframe with reading study results in inDFtype format

inDFtype
A string indicating the format type of the input MRMC data frame

outDFtype
A string indicating the format type of the output MRMC data frame.

readers
A character array holding the column names (readerIDs) corresponding to the data from different readers

Details

MRMC data frames contain scores from readers, cases, and sometimes modalities. This packages deals with (currently) two MRMC data frame formats. These are the formats:

matrixMode

For this format, each row contains all the information for a case, including the reader study result for several readers, ground truth and other information as seperate columns. This mode can only hold data from one modality.

listMode

For this format, each row contains the data for one observation. The columns specify the readerID, caseID, score, ground truth, and other information if there is any. This mode can hold data from multiple modalities.

Value

An MRMC dataframe with reading study results in outDFtype format

[Package iMRMC version 1.2.3 ]