createGroups {iMRMC}	R Documentation

Assign a group label to items in a vector

Description

Assign a group label to items in a vector

Usage


createGroups(items, nG)

Arguments

items
A vector of items

nG
The number of groups

Value

A data frame containing the items and their group labels

Examples


x <- paste("item", 1:10, sep = "")
df <- createGroups(x, 3)
print(df)


[Package iMRMC version 1.2.0 ]