createIMRMCdf {iMRMC}	R Documentation

Convert a data frame with all needed factors to doIMRMC formatted data frame

Description

Convert a data frame with all needed factors to doIMRMC formatted data frame

Usage


createIMRMCdf(dFrame, keyColumns = list(readerID = "readerID", caseID =
  "caseID", modalityID = "modalityID", score = "score", truth = "truth"),
  truePositiveFactor = "cancer")

Arguments

dFrame
This data frame includes columns for readerID, caseID, modalityID, score, and truth. These columns are not expected to be named as such and other columns may exist.

keyColumns
This list identifies the column names of the data frame to be used for the analysis. list(readerID = "***", caseID = "***", modalityID = "***", score = "***", truth="***")

truePositiveFactor
The true positive label, such as "cancer" or "1"

Value

output a doIMRMC formatted data frame: rows for truth and rows for data

[Package iMRMC version 1.2.0 ]