extractPairedComparisonsWRBM {iMRMC}	R Documentation

Extract within-reader between-modality pairs of scores

Description

Extract within-reader between-modality pairs of scores

Usage


extractPairedComparisonsWRBM(data0, modalities = "testA",
  keyColumns = list(readerID = "readerID", caseID = "caseID", modalityID
  = "modalityID", score = "score"))

Arguments

data0
This data frame includes columns for readerID, caseID, modalityID, score.

modalities
The modalities (testA, testB) for the scores to be paired

keyColumns
This list identifies the column names of the data frame to be used for the analysis. list(readerID = "***", caseID = "***", modalityID = "***", score = "***", truth="***")

Value

A data frame of all paired observations. Each observation comes from a one reader evaluating a case in two modalities The first column corresponds to one reader evaluating the case in testA. The second column corresonds to the same reader evaluating the case in testB.

[Package iMRMC version 1.2.0 ]