la.anova	R Documentation

MRMC Analysis of Limits of Agreement using ANOVA

Description

These two functions calculate two types of Limits of Agreement using ANOVA: Within-Reader Between-Modality(WRBM) and Between-Reader Between-Modality(BRBM). The 95% confidence interval of the mean difference is also provided.

Usage

laWRBM.anova(
  df,
  modalitiesToCompare = c("testA", "testB"),
  keyColumns = c("readerID", "caseID", "modalityID", "score"),
  if.aov = TRUE
)

laBRBM.anova(
  df,
  modalitiesToCompare = c("testA", "testB"),
  keyColumns = c("readerID", "caseID", "modalityID", "score"),
  if.aov = TRUE
)

Arguments

df
Data frame of observations, one per row. Columns identify random effects, fixed effects, and the observation. Namely,

readerID

The factor corresponding to the different readers in the study. The readerID is treated as a random effect.

caseID

The factor corresponding to the different cases in the study. The caseID is treated as a random effect.

modalityID

The factor corresponding to the different modalities in the study. The modalityID is treated as a fixed effect.

score

The score given by the reader to the case for the modality indicated.

modalitiesToCompare
The factors identifying the modalities to compare. It should be at length 2. Default modalitiesToCompare = c("testA","testB")

keyColumns
Identify the factors corresponding to the readerID, caseID, modalityID, and score (or alternative random and fixed effects). Default keyColumns = c("readerID", "caseID", "modalityID", "score")

if.aov
Boolean value to determine whether using aov function to do ANOVA. Default if.aov = TRUE

Details

Suppose the score from reader j for case k under modality i isX_{ijk}, then the difference score from the same reader for the same cases under two different modalities is Y_{jk} = X_{1jk} - X_{2jk}.

laWRBM use two-way random effect ANOVA to analyze the difference scores Y_{jk}. The model is Y_{jk}=μ + R_j + C_k + ε_{jk}, where R_j and C_k are random effects for readers and cases. The variance of mean and individual observation is expressed as the linear combination of the MS given by ANOVA.

laBRBM use three-way mixed effect ANOVA to analyze the scores X_{ijk}. The model is given by X_{ijk}=μ + R_j + C_k + m_i + RC_{jk} + mR_{ij} + mC_{ik} + ε_{ijk}, where R_j and C_k are random effects for readers and cases and m_i is a fixed effect for modality. The variance of mean and individual observation is expressed as the linear combination of the MS given by ANOVA.

Value

A dataframe with one row. Each column is as following:

meanDiff

The mean of difference score.

var.MeanDiff

The variance of mean difference score

var.1obs

The variance of a single WRBM/BRBM difference score

ci95meanDiff.bot

Lower bound of 95% CI for the mean difference score. meanDiff+ 1.96*sqrt(var.MeanDiff)

ci95meanDiff.top

Upper bound of 95% CI for the mean difference score. meanDiff- 1.96*sqrt(var.MeanDiff)

la.bot

Lower bound of WRBM/BRBM Limits of Agreement. meanDiff+2*sqrt(var.1obs)

la.top

Upper bound of WRBM/BRBM Limits of Agreement. meanDiff-2*sqrt(var.1obs)

The two function shows the same 95% CI for the mean difference score, but difference Limits of Agreements.

Examples

library(iMRMC)
# Create an MRMC data frame
# Refer to Gallas2014_J-Med-Img_v1p031006
simRoeMetz.config <- sim.gRoeMetz.config()

# Simulate data
df.MRMC <- sim.gRoeMetz(simRoeMetz.config)

# Compute Limits of Agreement
laWRBM_result <- laWRBM.anova(df.MRMC)
print(laWRBM_result)
laBRBM_result <- laBRBM.anova(df.MRMC)
print(laBRBM_result)