laBRBM {iMRMC}	R Documentation

MRMC analysis of between-reader between-modality limits of agreement

Description

The core analysis is done by ustat11 with the identity kernel (kernelFlag = 1).

Usage


laBRBM(df, modalitiesToCompare, keyColumns = c("readerID", "caseID",
  "modalityID", "score"))

Arguments

df
Data frame of observations, one per row. Columns identify random effects, fixed effects, and the observation. Namely,

readerID: The factor corresponding to the different readers in the study. The readerID is treated as a random effect.

caseID: The factor corresponding to the different cases in the study. The caseID is treated as a random effect.

modalityID: The factor corresponding to the different modalities in the study. The modalityID is treated as a fixed effect.

score: The score given by the reader to the case for the modality indicated.

modalitiesToCompare
The factors identifying the modalities to compare.

keyColumns
Identify the factors corresponding to the readerID, caseID, modalityID, and score (or alternative random and fixed effects).

[Package iMRMC version 1.2.0 ]