renameCol {iMRMC}	R Documentation

Rename a data frame column name or a list object name

Description

Rename a data frame column name or a list object name

Usage


renameCol(df, oldColName, newColName)

Arguments

df
A data frame

oldColName
Old column name

newColName
New column name

Value

the data frame with the updated column name

[Package iMRMC version 1.2.0 ]