roeMetzConfigs {iMRMC}	R Documentation

roeMetzConfigs

Description

This is a data frame containing the configuration parameters used in Roe1997_Acad-Radiol_v4p298. Each row corresponds to one of the twelve configurations appearing in Table 1 of that paper in a format that can be the input to sim.gRoeMetz.

Details

The columns of this data frame are as follows

Experiment labels and size

modalityID.A: [chr] label modality A

modalityID.B: [chr] label modality B

nR: [num] number of readers

nC.neg: [num] number of signal-absent cases

nC.pos: [num] number of signal-present cases

There are six fixed effects:

mu.neg: [num] signal-absent (neg, global mean)

mu.pos: [num] signal-present (pos, global mean)

mu.Aneg: [num] modality A signal-absent (Aneg, modality effect)

mu.Bneg: [num] modality B signal-absent (Bneg, modality effect)

mu.Apos: [num] modality A signal-present (Apos, modality effect)

mu.Bpos: [num] modality B signal-present (Bpos, modality effect)

There are six random effects that are independent of modality

var_r.neg: [num] variance of random reader effect

var_c.neg: [num] variance of random case effect

var_rc.neg: [num] variance of random reader by case effect

var_r.pos: [num] variance of random reader effect

var_c.pos: [num] variance of random case effect

var_rc.pos: [num] variance of random reader by case effect

There are six random effects that are specific to modality A

var_r.Aneg: [num] variance of random reader effect

var_c.Aneg: [num] variance of random case effect

var_rc.Aneg: [num] variance of random reader by case effect

var_r.Apos: [num] variance of random reader effect

var_c.Apos: [num] variance of random case effect

var_rc.Apos: [num] variance of randome reader by case effect

There are six random effects that are specific to modality B

var_r.Bneg: [num] variance of random reader effect

var_c.Bneg: [num] variance of random case effect

var_rc.Bneg: [num] variance of random reader by case effect

var_r.Bpos: [num] variance of random reader effect

var_c.Bpos: [num] variance of random case effect

var_rc.Bpos: [num] variance of randome reader by case effect

[Package iMRMC version 1.2.0 ]