sim.NormalIG.Hierarchical {iMRMC}	R Documentation

Simulate an MRMC data set comparing two modalities by a hierarchical model

Description

This procedure simulates an MRMC data set for a MRMC agreement study comparing two modalities.It is a hierarchical model consists of two interaction terms: reader-case interaction and modality-reader-case-replicate interaction. Both the interaction terms are conditional normal distributed, with the case(-related) factor contributing to the conditional mean and the reader(-related) factor contributing to the conditional variance.

Usage

sim.NormalIG.Hierarchical(
  config,
  R = NULL,
  AR = NULL,
  BR = NULL,
  is.within = FALSE
)

Arguments

config
[list] of simulation parameters:

Experiment labels and size

modalityID: [vector] label modality A and B.

nR: [num] number of readers

nC: [num] number of cases

C_dist: [chr] distribution of the case. Default C_dist="normal"

Mean and fixed effects:

mu: [num] grand mean

tau_A: [num] modality A

tau_B: [num] modality B

Reader-case interaction term

sigma_C: [num] variance of case factor (if C_dist="normal")

a_C: [num] alpha for distribution of case (if C_dist="beta")

b_C: [num] beta for distribution of case (if C_dist="beta")

alpha_R: [num] shape parameter for reader

beta_R: [num] scale parameter for reader

Modality-reader-case-replicate interaction term for modality A

sigma_C.A: [num] variance of case factor (if C_dist="normal")

a_C.A: [num] alpha for distribution of case (if C_dist="beta")

b_C.A: [num] beta for distribution of case (if C_dist="beta")

alpha_R.A: [num] shape parameter for reader

beta_R.A: [num] scale parameter for reader

Modality-reader-case-replicate interaction term for modality B

sigma_C.B: [num] variance of case factor (if C_dist="normal")

a_C.B: [num] alpha for distribution of case (if C_dist="beta")

b_C.B: [num] beta for distribution of case (if C_dist="beta")

alpha_R.B: [num] shape parameter for reader

beta_R.B: [num] scale parameter for reader

Scales for the case related terms and interaction terms

C_scale: [num] weight for the case factor

RC_scale: [num] weight for the reader-case interaction term

tauC_scale: [num] weight for the modality-case term

tauRCE_scale: [num] weight for the modality-reader-case-replicate interaction term

R
[vector] fix the reader factor across different simulation. Default R = NULL

AR
[vector] fix the modality-reader interaction. Default AR = NULL

BR
[vector] fix the modality-reader interaction. Default BR = NULL

is.within
[bol] whether the data are within-modality (AR==BR). Default is.within=FALSE

Details

The model is as the following structure: X.ijkl = mu + m.i + RC.jk + mRCE.ijkl

mu = grand mean

m.i = modalities (levels: A and B)

RC.jk|R.j,C.k ~ N(C.k, R.j) reader-case interaction term

mRCE.ijkl|mR.ij,mC.ik ~ N(mC.ik, mR.ij) modality-reader-case-replicate term

C.k and mC.ik are Normal/beta distributed

R.j and mR.ij are Inverse-Gamma distributed

Value

df [data.frame] with nR x nC x 2 rows including

readerID: [Factor] w/ nR levels "reader1", "reader2", ...

caseID: [Factor] w/ nC levels "case1", "case2", ...

modalityID: [Factor] w/ 1 level config$modalityID

score: [num] reader score

[Package iMRMC version 1.2.3 ]