sim.gRoeMetz {iMRMC}	R Documentation

Simulate an MRMC data set of an ROC experiment comparing two modalities

Description

This procedure simulates an MRMC data set of an ROC experiment comparing two modalities. It is based on Gallas2014_J-Med-Img_v1p031006, which generalizes of the model in Roe1997_Acad-Radiol_v4p298 and Roe1997_Acad-Radiol_v4p587. Specifically, it allows the variance components to depend on the truth and the modality. For the simpler Roe and Metz model, you can enter the smaller set of parameters into sim.gRoeMetz.config and get back the larger set of parameters and then used with this function.

Usage


sim.gRoeMetz(config)

Arguments

config
[list] of simulation parameters:

Experiment labels and size

modalityID.A: [chr] label modality A

modalityID.B: [chr] label modality B

nR: [num] number of readers

nC.neg: [num] number of signal-absent cases

nC.pos: [num] number of signal-present cases

There are six fixed effects:

mu.neg: [num] signal-absent (neg, global mean)

mu.pos: [num] signal-present (pos, global mean)

mu.Aneg: [num] modality A signal-absent (Aneg, modality effect)

mu.Bneg: [num] modality B signal-absent (Bneg, modality effect)

mu.Apos: [num] modality A signal-present (Apos, modality effect)

mu.Bpos: [num] modality B signal-present (Bpos, modality effect)

There are six random effects that are independent of modality

var_r.neg: [num] variance of random reader effect

var_c.neg: [num] variance of random case effect

var_rc.neg: [num] variance of random reader by case effect

var_r.pos: [num] variance of random reader effect

var_c.pos: [num] variance of random case effect

var_rc.pos: [num] variance of random reader by case effect

There are six random effects that are specific to modality A

var_r.Aneg: [num] variance of random reader effect

var_c.Aneg: [num] variance of random case effect

var_rc.Aneg: [num] variance of random reader by case effect

var_r.Apos: [num] variance of random reader effect

var_c.Apos: [num] variance of random case effect

var_rc.Apos: [num] variance of randome reader by case effect

There are six random effects that are specific to modality B

var_r.Bneg: [num] variance of random reader effect

var_c.Bneg: [num] variance of random case effect

var_rc.Bneg: [num] variance of random reader by case effect

var_r.Bpos: [num] variance of random reader effect

var_c.Bpos: [num] variance of random case effect

var_rc.Bpos: [num] variance of randome reader by case effect

Details

The simulation is a linear model with six fixed effects related to modality and truth and 18 normally distributed independent random effects for readers, cases, and the interaction between the two. Here is the linear model:

L.mrct = mu.t + mu.mt
+ reader.rt + case.ct + readerXcase.rct
+ modalityXreader.mrt + modalityXcase.mct + modalityXreaderXcase.mrct

m=modality (levels: A and b)

t=truth (levels: neg and Pos)

mu.t is the global mean for t=neg and t=pos cases

mu.mt is the modality specific fixed effects for t=neg and t=pos cases

the remaining terms are the random effects: all independent normal random variables

Value

dFrame.imrmc [data.frame] with (nC.neg + nC.pos)*(nR+1) rows including

readerID: [Factor] w/ nR levels "reader1", "reader2", ...

caseID: [Factor] w/ nC levels "case1", "case2", ...

modalityID: [Factor] w/ 1 level config$modalityID

score: [num] reader score

Note that the first nC.neg + nC.pos rows specify the truth labels for each case. For these rows, the readerID must be "truth" or "-1" and the score must be 0 for negative cases and 1 for postive cases.

[Package iMRMC version 1.2.0 ]