sim.new.Hierarchical.config	R Documentation

Create a configuration object for the sim.new.Hierarchical program

Description

#' @description This function creates a configuration object for the Hierarchical simulation model to be used as input for the sim.new.Hierarchical program.

Usage

sim.new.Hierarchical.config(
  nR = 5,
  nC = 100,
  modalityID = c("testA", "testA*"),
  C_dist = "normal",
  mu = 0,
  tau_A = 0,
  tau_B = 0,
  alpha_R = 10,
  beta_R = 1,
  sigma_C = 1,
  a_C = 0.8,
  b_C = 3,
  sigma_tauC = 1,
  alpha_tauR = 10,
  beta_tauR = 1,
  C_scale = 1,
  RC_scale = 1,
  tauC_scale = 1,
  tauRCE_scale = 1
)

Arguments

nR
[num] Number of readers. Default nR = 5

nC
[num] Number of cases. Default nC = 100

modalityID
[vector] List of modalityID. Default modalityID = c("testA", "testA*")

C_dist
[chr] Distribution of the case. Default C_dist="normal"

mu
[num] grand mean. Default mu = 0

tau_A
[num] modality A effect. Default tau_A = 0

tau_B
[num] modality B effect. Default tau_B = 0

alpha_R
[num] shape parameter for reader. Default alpha_R = 10

beta_R
[num] scale parameter for reader. Default beta = 1

sigma_C
[num] variance of case factor (if C_dist="normal"). Default sigma_C = 1

a_C
[num] alpha for distribution of case (if C_dist="beta"). Default a_C = 0.8

b_C
[num] beta for distribution of case (if C_dist="beta"). Default b_C = 3

sigma_tauC
[num] variance of modality-case (if C_dist="normal"). Default sigma_tauC = 1

alpha_tauR
[num] shape parameter for modality-reader. Default alpha_tauR = 10

beta_tauR
[num] scale parameter for modality-reader. Default beta_tauR = 1

C_scale
[num] weight for the case factor. Default C_scale = 1

RC_scale
[num] weight for the reader-case interaction term. Default RC_scale = 1

tauC_scale
[num] weight for the modality-case term. Default tauC_scale = 1

tauRCE_scale
[num] weight for the modality-reader-case-replicate interaction term. Default tauRCE_scale = 1

Details

If no arguments, this function returns a default simulation configuration for sim.new.Hierarchical

Value

config [list] Refer to the sim.new.Hierarchical input variable