simMRMC {iMRMC}	R Documentation

Simulate an MRMC data set ####

Description

This program simulates observations from one set of readers scoring one set of cases. It produces one modality and one truth state of ROC data following Roe1997_Acad-Radiol_v4p298 and Roe1997_Acad-Radiol_v4p587. In order to produce an entire ROC data set, please use sim.gRoeMetz.

Usage


simMRMC(simMRMC.config)

Arguments

simMRMC.config
[list] of simulation parameters:

modalityID [chr] label modalityID

readerIDs [factor] the ID of each reader

caseIDs [factor] the ID of each case

mu [num] mean

var_r [num] variance of random reader effect

var_c [num] variance of random case effect

var_rc [num] variance of random reader by case effect

Details

The simulation is a linear model with one fixed effect and three normally distributed independent random effects corresponding to readers, cases, and an interaction between the two.

L.rc = mu + readerEffect.r + caseEffect.c + readerXcaseEffect.rc

Value

L [data.frame] with nC*nR rows of 4 variables

L$modalityID [factor] determined by input modalityID

L$readerID [factor] determined by input readerIDs

L$caseID [factor] determined by input caseIDs

L$score [num] R.r + C.c + RC.rc

r = 1,2,...,nR

c = 1,2,...,nC

R.r ~ N(0,var_r)

C.c ~ N(0,var_c)

RC.rc ~ N(0,var_rc)

[Package iMRMC version 1.2.0 ]