successDFtoROCdf {iMRMC}	R Documentation

Convert an MRMC data frame of successes to one formatted for doIMRMC

Description

Convert an MRMC data frame of successes to one formatted for doIMRMC

Usage


successDFtoROCdf(df)

Arguments

df
Each row contains a success observation for one reader evaluating one case

Value

data frame ready for doIMRMC

[Package iMRMC version 1.2.0 ]