undoIMRMCdf {iMRMC}	R Documentation

Convert a doIMRMC formatted data frame to a standard data frame with all factors.

Description

Convert a doIMRMC formatted data frame to a standard data frame with all factors.

Usage


undoIMRMCdf(df.MRMC)

Arguments

df.MRMC
This data frame includes columns for readerID, caseID, modalityID, score. Each row is a reader x case x modality observation from the study In addition to observations from the study, this data frame requires rows specifying the truth for each caseID. For truth specifications, the readerID needs to equal "truth" or "-1", modalityID can be anything ("truth" is a good choice), and score should be 0 for signal-absent normal case, 1 for signal-present disease case.

Details

Delete rows specifying truth and put the truth information on every row.

Value

output a data frame with columns readerID, caseID, modalityID, score, truth

[Package iMRMC version 1.2.0 ]