convertDF {iMRMC}	R Documentation

Convert a data frame from matrix format to list format

Description

Convert a data frame from matrix format to list format

Usage


convertDF(inDF, inDFtype, outDFtype, readers, nameTruth)

Arguments

inDF
A dataframe with reading study result in inDFtype format

inDFtype
A string indicating the format type of the input data frame

outDFtype
A string indicating the format type of the output data frame.

readers
A list of strings presenting the readerIDs

nameTruth
NULL

Details

matrixWithTruth indicating the format type that each row contains all the information for a case, including the reader study result for several readers, ground truth and other information as seperate columns;

listWithTruth indicating the format type that includes columns for readerID, caseID, score, ground truth, and other information if there is any.

Value

A dataframe with reading study result in outDFtype format

[Package iMRMC version 1.2.0 ]